+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Added
-* Add `WDIO_IGNORE_COMPARISON_RESULTS` env to disabled assertions on screenshot comparison results 
+* Add `WDIO_IGNORE_COMPARISON_RESULTS` env to disable assertions on screenshot comparison results 
 
 5.18.1 - (December 18, 2019)
 ----------